@@ -10,54 +10,54 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.2.12",
-    "@angular/common": "^16.2.0",
-    "@angular/compiler": "^16.2.0",
-    "@angular/core": "^16.2.0",
-    "@angular/forms": "^16.2.12",
-    "@angular/platform-browser": "^16.2.0",
-    "@angular/platform-browser-dynamic": "^16.2.0",
-    "@angular/router": "^16.2.0",
-    "@ngx-formly/core": "^6.3.0",
-    "@orama/orama": "^2.0.0-beta.8",
-    "@orama/stemmers": "^2.0.0-beta.9",
-    "@swc/core": "^1.3.100",
+    "@angular/animations": "^18.1.0",
+    "@angular/common": "^18.1.0",
+    "@angular/compiler": "^18.1.0",
+    "@angular/core": "^18.1.0",
+    "@angular/forms": "^18.1.0",
+    "@angular/platform-browser": "^18.1.0",
+    "@angular/platform-browser-dynamic": "^18.1.0",
+    "@angular/router": "^18.1.0",
+    "@ngx-formly/core": "^6.3.6",
+    "@orama/orama": "^2.0.21",
+    "@orama/stemmers": "^2.0.21",
+    "@swc/core": "^1.6.13",
     "assert": "^2.1.0",
     "constants": "^0.0.2",
     "crypto-browserify": "^3.12.0",
-    "express": "^4.18.2",
+    "express": "^4.19.2",
     "https-browserify": "^1.0.0",
-    "ini": "^4.1.1",
-    "mssql": "^10.0.1",
-    "ngx-cookie-service": "^16.0.2",
-    "ngx-toastr": "^18.0.0",
+    "ngx-cookie-service": "^18.0.0",
+    "ngx-toastr": "^19.0.0",
     "orama": "^2.0.6",
     "os-browserify": "^0.3.0",
-    "rxjs": "~7.8.0",
+    "rxjs": "~7.8.1",
     "simple-node-logger": "^21.8.12",
     "stream-browserify": "^3.0.0",
     "stream-http": "^3.2.0",
-    "tslib": "^2.3.0",
+    "tslib": "^2.6.3",
     "wcs-angular": "^4.1.0",
     "wcs-core": "^4.2.0",
     "wcs-formly": "^4.2.0",
-    "zone.js": "~0.13.0"
+    "zone.js": "~0.14.7"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.2.9",
-    "@angular/cli": "^16.2.9",
-    "@angular/compiler-cli": "^16.2.0",
-    "@ngx-formly/schematics": "^6.3.0",
-    "@types/cookie-parser": "^1.4.6",
-    "@types/jasmine": "~4.3.0",
+    "@angular-devkit/build-angular": "^18.1.0",
+    "@angular/cli": "^18.1.0",
+    "@angular/compiler-cli": "^18.1.0",
+    "@ngx-formly/schematics": "^6.3.6",
+    "@types/cookie-parser": "^1.4.7",
+    "@types/jasmine": "~5.1.4",
+    "@types/lodash": "^4.17.7",
     "@types/mssql": "^9.1.5",
-    "electron": "^27.0.4",
-    "jasmine-core": "~4.6.0",
-    "karma": "~6.4.0",
+    "electron": "^31.2.1",
+    "jasmine-core": "~5.1.2",
+    "karma": "~6.4.3",
     "karma-chrome-launcher": "~3.2.0",
-    "karma-coverage": "~2.2.0",
+    "karma-coverage": "~2.2.1",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "typescript": "~5.1.3"
+    "lodash": "^4.17.21",
+    "typescript": "~5.5.3"
   }
 }
