@@ -1,18 +1,25 @@
 import {App} from "~/server";
 import {Request, Response} from "express";
-import {AuthResponseType, RequestType, UserType} from "~/types/types";
+import {AuthUsername, RequestType, RoleType, UserType} from "~/types/types";
 import {VerifyErrors} from "jsonwebtoken";
-const jwt = require("jsonwebtoken");
-
+import {API_Error} from "~/types/errors";
+import * as jwt from "jsonwebtoken";
+import bcrypt from "bcryptjs";
 
 export class AuthModule {
 
   MAIN_TOKEN: string
   App: App
+  Users: Map<string, UserType> = new Map()
 
   constructor(mainClass: App) {
     this.App = mainClass
     this.MAIN_TOKEN = mainClass.config.auth.mainToken
+
+    // Create a HashMap of users
+    this.App.config.users.forEach((user) => {
+      this.Users.set(user.username, user)
+    })
   }
 
   createUserToken(username: string): string {
@@ -20,77 +27,153 @@ export class AuthModule {
     return userToken
   }
 
-  checkUserPassword(req: Request, res: Response): AuthResponseType {
+  checkUserPassword(req: Request, res: Response): UserType {
     let reqObject: RequestType = req.body as RequestType
-    let authResponseObject: AuthResponseType = new AuthResponseType()
     if (!reqObject || !reqObject.data.username || !reqObject.data.password) {
-      authResponseObject.code = 400
-      authResponseObject.message = "Il manque des éléments dans ta requête..."
-      this.App.sendResponse(res, undefined, authResponseObject)
-      return authResponseObject
+      throw new API_Error('BAD_REQUEST', 'Il manque des éléments dans ta requête...', {code: 401})
     }
-    let userFind: UserType | undefined = this.App.config.adminUsers.find((user: UserType) => user.username == reqObject.data.username && user.password == reqObject.data.password)
+    let username = reqObject.data.username
+    let password = reqObject.data.password
+    let userFind: UserType | undefined = this.Users.get(username)
     if (!userFind) {
-      authResponseObject.code = 401
-      authResponseObject.message = "Les identifiants fournis sont incorrects..."
-      this.App.sendResponse(res, undefined, authResponseObject)
-      return authResponseObject
+      throw new API_Error('BAD_CREDENTIALS', "L'utilisateur n'a pas été trouvé...", {code: 401})
     }
-    authResponseObject.username = reqObject.data.username
-    return authResponseObject
+    let match = bcrypt.compareSync(password, userFind.password)
+    if (!match) {
+      throw new API_Error('BAD_CREDENTIALS', "Mauvais mot de passe !", {code: 401})
+    }
+    return userFind
   }
 
-  checkUserToken(req: Request, res: Response): AuthResponseType { // true if authentificated and false if not
+  checkUserTokenFromPOST(req: Request, res: Response): UserType { // true if authentificated and false if not
     let reqObject: RequestType = req.body
-    let authResponseObject: AuthResponseType = new AuthResponseType()
-
     if (!reqObject) {
-      authResponseObject.code = 401
-      authResponseObject.message = "Il manque du contenu dans la requête"
-      this.App.sendResponse(res, undefined, authResponseObject)
-      return authResponseObject
+      throw new API_Error('BAD_REQUEST', 'Il manque du contenu dans la requête', {code: 401})
     }
     try {
-      let token = reqObject.data.token
+      let token = reqObject.data.token // TODO : Améliorer ce check de variable --> faire un if === undefined à la place
     } catch(err) {
-      authResponseObject.code = 401
-      authResponseObject.message = "Il manque le token dans votre requête..."
-      this.App.sendResponse(res, undefined, authResponseObject)
-      return authResponseObject
+      throw new API_Error('REQUEST_VALUES_MISSING', 'Il manque le token dans votre requête...', {code: 401})
     }
 
-    let username: undefined | string = undefined
     try {
       let token = ""
       if (reqObject.data.token) {token = reqObject.data.token}
       if (reqObject.token) {token = reqObject.token}
-      username = jwt.verify(token, this.App.config.auth.mainToken).username
+      let username = (jwt.verify(token, this.MAIN_TOKEN) as any).username
+
+      // Si pas d'erreur au dessus, jwt valide
+      let user = this.Users.get(username)
+      if (!user) {
+        throw new API_Error('BAD_CREDENTIALS', 'Utilisateur non trouvé dans la base de donnée')
+      }
+      return user
+
     } catch (err) {
       let errObject: VerifyErrors = (err as VerifyErrors)
       if (errObject?.name == "TokenExpiredError") {
-        authResponseObject.code = 401
-        authResponseObject.message = "Ton token a expiré"
+        throw new API_Error('EXPIRED_TOKEN', "Ton token a expiré", {code: 401})
       } else {
-        authResponseObject.code = 401
-        authResponseObject.message = "Ton token est invalide"
+        throw new API_Error('INVALID_TOKEN', "Ton token est invalide", {code: 401})
       }
     }
 
-    if (authResponseObject.code != 200 || !username) {
-      this.App.sendResponse(res, undefined, authResponseObject)
-      return authResponseObject
-    } else {
-      // Ajouter le temps de dernière connexion
-      this.App.config.adminUsers.some((user: UserType) => {
-        if (user.username == username as any) {
-          user.lastConnect = new Date().getTime()
-          return
-        }
-      })
-      this.App.writeConfig()
-      authResponseObject.username = username
-      return authResponseObject
+  }
+
+  checkUserTokenFromGET(req: Request, res: Response): UserType {
+    let reqHeaders = req.headers
+
+    if (reqHeaders.token == undefined) {
+      throw new API_Error('REQUEST_VALUES_MISSING', 'Il manque le token dans votre requête...', {code: 401})
     }
+
+    try {
+      let token: string = (reqHeaders.token as string)
+      let username = (jwt.verify(token, this.MAIN_TOKEN) as any).username
+
+      // Si pas d'erreur au dessus, jwt valide
+      let user = this.Users.get(username)
+      if (!user) {
+        throw new API_Error('BAD_CREDENTIALS', 'Utilisateur non trouvé dans la base de donnée')
+      }
+      return user
+
+    } catch (err) {
+      let errObject: VerifyErrors = (err as VerifyErrors)
+      if (errObject?.name == "TokenExpiredError") {
+        throw new API_Error('EXPIRED_TOKEN', "Ton token a expiré", {code: 401})
+      } else {
+        throw new API_Error('INVALID_TOKEN', "Ton token est invalide", {code: 401})
+      }
+    }
+  }
+
+  createUser(username: string, password: string, role: RoleType): void {
+    if (this.Users.has(username)) {
+      throw new API_Error('USER_ALREADY_EXISTS', "L'utilistateur souhaité existe déjà...", {code: 409})
+    }
+    if (password == '' || username == '') {
+      throw new API_Error('BAD_CREDENTIALS', "L'utilistateur ou mot de passe ne peuvent pas être vides...", {code: 401})
+    }
+    let hashedPassword = this.createHashUserPassword(password)
+    let userObject: UserType = {
+      username: username,
+      password: hashedPassword,
+      role: role
+    }
+    this.writeUser(username, userObject)
+  }
+
+  checkRole(wantedRole: RoleType, user: UserType): void {
+    let wantedRoleWeight = this.App.config.roles[wantedRole]
+    let userRoleWeight = this.App.config.roles[user.role]
+    if (userRoleWeight < wantedRoleWeight) {
+      throw new API_Error('USER_ROLE_INSUFFICIENT', "Vous n'avez pas le droit d'accéder à cette ressource", {code: 403})
+    }
+  }
+
+  createHashUserPassword(password: string): string {
+    return bcrypt.hashSync(password, 10)
+  }
+
+  writeUser(username: string, userObject: UserType): void {
+    if (!this.Users.has(username)) {
+      this.App.config.users.push(userObject)
+      this.App.writeConfig()
+    } else {
+      let foundUserObject = this.App.config.users.find((user) => user.username === username)
+      if (!foundUserObject) {
+        throw new API_Error('BAD_CREDENTIALS', 'Strange to have a user not found in the user DB here...', {code: 500})
+      }
+      let userIndex = this.App.config.users.indexOf(foundUserObject)
+      this.App.config.users.splice(userIndex, 1)
+      this.App.config.users.push(userObject)
+      this.App.writeConfig()
+    }
+    this.Users.set(username, userObject)
+  }
+
+  deleteUser(username: string) {
+    if (!this.Users.has(username)) {
+      throw new API_Error('BAD_CREDENTIALS', "L'utilisateur n'a pas été trouvé...", {code: 404})
+    }
+    this.Users.delete(username)
+    let foundUserObject = this.App.config.users.find((user) => user.username === username)
+    if (!foundUserObject) {
+      throw new API_Error('BAD_CREDENTIALS', "L'utilisateur demandé n'a pas été trouvé...", {code: 404})
+    }
+    let userIndex = this.App.config.users.indexOf(foundUserObject)
+    this.App.config.users.splice(userIndex, 1)
+    this.App.writeConfig()
+  }
+
+  addUserConnectDate(username: string): void {
+    let userObject = this.Users.get(username)
+    if (!userObject) {
+      throw new API_Error('BAD_CREDENTIALS', "User not found...", {code: 404})
+    }
+    userObject.lastConnect = new Date().getTime()
+    this.writeUser(username, userObject)
   }
 
 }
