@@ -1,244 +1,180 @@
 import {App} from "~/server";
-import {DBconfigType, RawDataTables, TableObjectType} from "~/types/types";
-import sql, {IRecordSet} from "mssql";
+import {DBconfigType, NOT_TO_DATE, TableName, TableObjectType, UP_TO_DATE} from "~/types/types";
 import {EventEmitter} from 'node:events';
-
+import * as console from "node:console";
+import {Client, connect, DatabaseError, ResultRecord} from "ts-postgres";
+import {API_Error} from "~/types/errors";
+import * as process from "node:process";
+import _ from "lodash";
 
 export class DataModule {
 
   App: App
   DBconfig: DBconfigType
-  SQLpool: sql.ConnectionPool
-  $dataReady: EventEmitter
-  rawTablesData: Map<string, IRecordSet<any>>
-  purifiedTablesData: Map<string, any>
-  alltablePurified: TableObjectType[]
-  lastRefresh: Date
+  lastRefresh: Date = new Date(2000,1,1,0,0,0)
+  dbClient: Client
+  dbConnected: boolean = false
+  dbEvent: EventEmitter = new EventEmitter
+
+  allTables: Map<string, ResultRecord> = new Map<string, ResultRecord>
 
   constructor(mainClass: App) {
     this.App = mainClass
-    this.$dataReady = new EventEmitter()
-    this.rawTablesData = new Map<string, IRecordSet<any>>()
-    this.purifiedTablesData = new Map<string, any>()
-    this.alltablePurified = []
-    this.DBconfig = {
-      user:  mainClass.config.database.user,
-      password:  mainClass.config.database.password,
-      server:  mainClass.config.database.server,
-      database:  mainClass.config.database.database,
-      port:  mainClass.config.database.port,
-      options: {
-        trustedConnection: true,
-        encrypt: false,
-        enableArithAbort: true,
-        trustServerCertificate: true,
-      }
-    }
+    this.DBconfig = mainClass.config.database
 
-    this.connectToServer().catch(() => process.exit(2)).then(r => {
-      this.updateAllTable() // It will emit a $dataReadyEvent 'finished' when all tables has been feteched
+    this.dbEvent.on('connected', () => {
+      console.log('Connected to DB')
+      this.cacheData()
     })
 
-    this.$dataReady.on('dataFetched', () => {
-      this.purifyTables()
-      this.$dataReady.emit('finished')
+    this.dbEvent.on('refreshed', () => {
+      console.log('Finished refreshing')
+      this.lastRefresh = new Date()
     })
+
+    this.connectToServer()
+
+    setInterval(() => {this.cacheData()}, this.App.config.data.refreshRate * 1000)
 
   }
 
   async connectToServer() {
-    this.SQLpool = await sql.connect(this.DBconfig).catch(() => {
-      console.error("Erreur lors de la connexion au serveur")
+    console.log("Connecting to DB")
+    try {
+      this.dbClient = await connect({
+        host: this.DBconfig.server,
+        port: this.DBconfig.port,
+        user: this.DBconfig.user,
+        password: this.DBconfig.password,
+        database: this.DBconfig.database});
+      this.dbConnected = true;
+      this.dbEvent.emit('connected');
+      this.dbClient.on('end', () => {
+        this.dbConnected = false
+        this.dbEvent.emit('disconnected')
+      })
+    } catch (error) {
+      console.error(error)
+      console.log("Stopping the server...")
       process.exit(2)
-    })
-  }
-
-  async getTable(table: string, where?: {id: string, value: string}[]) {
-    let queryString = `SELECT * FROM ${table}`
-    // Add all the elements of where in the query
-    if (where) {
-      queryString += ' WHERE'
-      for (let i = 0; i < where.length; i++) {
-        queryString += ` ${where[i].id}='${where[i].value}'`
-        if (i < (where.length - 1)) {
-          queryString += ' AND'
-        }
-      }
-    }
-    queryString += ';'
-    // Make the request
-    try {
-      let results = (await this.SQLpool.request().query(queryString)).recordset
-      if (!results) {
-        console.error(`La liste de résultats de la requête ${queryString} est vide...`)
-        return null // TODO : See if there is not a better way to return this Promise in case of a error instead of returning a null...
-      }
-      return await results
-    } catch (error) {
-      console.error(`Une erreur est survenue sur la requête ${queryString}`, error)
-      return null // TODO : See if there is not a better way to return this Promise in case of a error instead of returning a null...
     }
   }
 
-  async addDocToDB(document: any) {
-    let queryString = 'INSERT INTO documents'
-    try {
-      let keysSQL = []
-      for (let element of Object.keys(document)) {
-        if (element != "id") {
-          keysSQL.push(element)
-        }
-      }
-      queryString += ` (${keysSQL.join(', ')})`
-      let valuesSQL = []
-      for (let element of Object.keys(document)) {
-        if (document[element] && element != "id") {
-          if (element == 'engin_type') {
-            valuesSQL.push(`'["${(document[element] as []).join('","')}"]'`)
-          } else {
-            valuesSQL.push(`'${document[element]}'`)
-          }
-        } else if (document[element] == null && element != "id") {
-          valuesSQL.push(`NULL`)
-        }
-      }
-      queryString += ` VALUES (${valuesSQL.join(', ')})`
-    } catch {
-      throw new Error("")
-    }
-    // Make the request
-    try {
-      let results = (await this.SQLpool.request().query(queryString)).recordset
-      console.log("SQL Query : ", queryString)
-      return await results
-    } catch (error) {
-      let errorMsg = `Une erreur est survenue sur la requête : ${queryString}`
-      console.error(errorMsg, error)
-      throw new Error(errorMsg)
-    }
-  }
+  async getAllTables() {
+    let allTableFetched = true
 
-  async modDocFromDB(id: number, document: any) {
-    let queryString = 'UPDATE documents SET '
-    try {
-      let valuesSQL = []
-      for (let element of Object.keys(document)) {
-        if (document[element] && element != 'id') {
-          if (element == 'engin_type') {
-            valuesSQL.push(`${element} = '["${(document[element] as []).join('","')}"]'`)
-          } else {
-            valuesSQL.push(`${element} = '${document[element]}'`)
-          }
-        } else if (document[element] == null && element != 'id') {
-          valuesSQL.push(`${element} = NULL`)
-        }
-      }
-      queryString += valuesSQL.join(', ')
-      queryString += ` WHERE id=${id}`
-    } catch (error) {
-      let errorMsg = `Une erreur est survenue sur la génération de la query SQL : ${queryString}`
-      console.error(errorMsg, error)
-      throw new Error(errorMsg)
-    }
-    // Make the request
-    try {
-      let results = (await this.SQLpool.request().query(queryString)).recordset
-      console.log("SQL Query : ", queryString)
-      return await results
-    } catch (error) {
-      let errorMsg = `Une erreur est survenue sur la query SQL : ${queryString}`
-      console.error(errorMsg, error)
-      throw new Error(errorMsg)
-    }
-  }
-
-  purifyTables() {
-
-    console.log("Purifying data")
-
-    this.alltablePurified = []
-    let timeRefresh = new Date().getTime()
-
-    // Technicentre Purify
-    let technicentreData = this.rawTablesData.get("technicentres")
-    let engins_technicentreData = this.rawTablesData.get("engins_technicentre")
-    if (technicentreData && engins_technicentreData) {
-      technicentreData = JSON.parse(JSON.stringify(technicentreData))
-      engins_technicentreData = JSON.parse(JSON.stringify(engins_technicentreData))
-      technicentreData?.forEach((technicentre: any) => {
-        technicentre.engins = []
-        engins_technicentreData?.filter((engin: any) => engin.technicentre == technicentre.technicentre).forEach((engin: any) => technicentre.engins.push({engin: engin.engin, engin_type:engin.engin_type, engin_numero: engin.num_engin}))
+    const tablePromises = this.App.config.data.allTableResponse.map(async (tableName: string) => {
+      const tableObject = await this.getTable(tableName).catch(() => {
+        allTableFetched = false
       })
-      this.purifiedTablesData.set("technicentres", technicentreData)
-      this.alltablePurified.push({tableName: "technicentres", tableData: technicentreData, tableLastRefresh: timeRefresh})
-    }
+      return tableObject
+    });
 
-    // News Purify
-    let newsData = this.rawTablesData.get("news")
-    if (newsData) {
-      newsData = JSON.parse(JSON.stringify(newsData))
-      newsData?.forEach((item) => item['urls'] && typeof item['urls'] == "string"  ? item['urls'] = JSON.parse(item['urls']) : null) // Convert JSON stored as text in DB
-      this.purifiedTablesData.set("news", newsData)
-      this.alltablePurified.push({tableName: "news", tableData: newsData, tableLastRefresh: timeRefresh})
-    }
+    let response = await Promise.all(tablePromises);
 
-    // Engins Purify
-    let enginsData = this.rawTablesData.get("engins")
-    let enginsTypesData = this.rawTablesData.get("engins_types")
-    if (enginsData && enginsTypesData) {
-      let newEngins: any[] = []
-      enginsData?.forEach((engin) => {
-        let found_types: any[] = []
-        enginsTypesData?.forEach((engin_type) => engin_type.engin == engin.engin ? found_types.push(engin_type.engin_type) : null)
-        newEngins.push({engin: engin.engin, types_engin: found_types, url_image_engin: engin.url_image_engin})
-      })
-      this.purifiedTablesData.set("engins", newEngins)
-      this.alltablePurified.push({tableName: "engins", tableData: newEngins, tableLastRefresh: timeRefresh})
+    response = response.filter(Boolean); // Filtrer les valeurs nulles ou indéfinies
+    if (response.length == 0 || !allTableFetched) {
+      throw new API_Error('API_INTERN_ERROR', "Cannot get tables from DB...")
     }
-
-    // Documents Purify
-    let documentsData = this.rawTablesData.get("documents")
-    if (documentsData) {
-      documentsData = JSON.parse(JSON.stringify(documentsData))
-      documentsData?.forEach((item) => item['engin_type'] && typeof item['engin_type'] == "string" ? item['engin_type'] = JSON.parse(item['engin_type']) : null) // Convert JSON stored as text in DB
-      this.purifiedTablesData.set("documents", documentsData)
-      this.alltablePurified.push({tableName: "documents", tableData: documentsData, tableLastRefresh: timeRefresh})
-    }
-
-    // Filters Purify
-    let filtersData = this.rawTablesData.get("filters")
-    if (filtersData) {
-      this.purifiedTablesData.set("filters", filtersData)
-      this.alltablePurified.push({tableName: "filters", tableData: filtersData, tableLastRefresh: timeRefresh})
-    }
-
-    console.log("Data purified")
-    console.log("PurifiedTables available : ", this.purifiedTablesData.keys())
+    return response
   }
 
-  async updateAllTable() {
-    console.log("Updating all data")
+  async getTable(tableName: string) {
+    try {
+      let query = await this.dbClient.query(`SELECT * FROM ${tableName}`)
+      let jsonData: ResultRecord = [...query]
+      let tableObject: TableObjectType = {tableName: tableName, tableData: jsonData, tableLastRefresh: new Date().getTime()}
+      return tableObject
+    } catch (error) {
+      if (error instanceof DatabaseError || error instanceof TypeError) {
+        console.error(`Erreur lors de la query table ${tableName} |`, error.message)
+        throw new API_Error('TABLE_NOT_FOUND', `La table ${tableName} n'a pas été trouvée`, {code: 404})
+      } else {
+        console.error(error)
+      }
+    }
+  }
 
-    let numOfTables = this.App.config.database.tables.length
-
-    this.App.config.database.tables.forEach((item, index) => {
-      this.getTable(item).then(record => {
-        if (record) {this.rawTablesData.set(item, record)} // Add the item to Map list
-
-        if (index+1 == numOfTables) { // Say that all tables has been fetched
-          console.log("Updated all data")
-          console.log("RawTables available : ", this.rawTablesData.keys())
-          this.lastRefresh = new Date() // Set the last refresh of all table to now
-          this.$dataReady.emit('dataFetched')
-        }
-      })
+  async cacheData() {
+    this.DBconfig.tables.forEach((table: string, id: number, arr) => {
+      this.getTable(table).then((result) => {
+        if (id+1 == arr.length) { this.dbEvent.emit('refreshed') } // Send that all tables has been refreshed if last
+        if (result) { this.allTables.set(table, result) }
+      }).catch((error) => {}) // TODO : Make something when there is a error...
     })
   }
 
   checkIfDataIsUpToDate() {
-    let timeDiffInMinutes = ((new Date().getTime() - this.lastRefresh.getTime())/1000)/60
-    if (timeDiffInMinutes >= this.App.config.data.refreshRate) {
-      this.updateAllTable() // Launch async funct. so the user will have the old data parsed and will need to do another request to have the new one
+    let timeDiffInSeconds = ((new Date().getTime() - this.lastRefresh.getTime())/1000)/3600
+    if (timeDiffInSeconds >= this.App.config.data.refreshLimit) {
+      this.cacheData()
+      return NOT_TO_DATE
+    } else {
+      return UP_TO_DATE
+    }
+  }
+
+  async createEntry(newObject: any, tableName: TableName): Promise<void> {
+    let keys = this.App.config.data.tablesNecessaryKeys.find(table => table.tableName == tableName)?.keys
+    if (!keys) {
+      throw new API_Error('API_INTERN_ERROR', `La config pour insérer un document dans la table ${tableName} n'a pas été trouvée...`)
+    }
+    if (!_.every(keys, _.partial(_.has, newObject))) {
+      throw new API_Error('REQUEST_VALUES_MISSING', 'Une / des clés de votre document est / sont manquantes dans votre requête...', {code: 401})
+    }
+
+    let values: string[] = []
+    keys.forEach((key: keyof typeof newObject) => {
+      try {
+        values = [...values, typeof newObject[key] == "string" ? newObject[key] as string : JSON.stringify(newObject[key])]
+      } catch (error) {
+        throw new API_Error('API_INTERN_ERROR', "Une erreur est survenue lors de la génération de l'objet...")
+      }
+    })
+
+    let keysString = '"' + keys.join("\", \"") + '"'
+    let valuesString = "'" + values.join("', '") + "'"
+
+    let checkQuery = `
+        SELECT EXISTS(
+            SELECT FROM "documents"
+            WHERE (engin = '${newObject.engin}' AND (name = '${newObject.name}' OR ref_main = '${newObject.ref_main}'))
+        )
+    `
+    let addQuery = `
+      INSERT INTO "documents" (${keysString})
+      VALUES (${valuesString});
+    `
+
+    let checkValue
+    try {
+      checkValue = await this.dbClient.query(checkQuery)
+    } catch (error) {
+      this.catchDataBaseErrror(error)
+    }
+
+    let checkBool = checkValue?.rows[0][0]
+    if (checkBool == true || checkBool == undefined) {
+      throw new API_Error('ENTRY_ALREADY_EXISTS', "Le document que vous souhaitez renseigner existe déjà...", {code: 409})
+    }
+
+    try {
+      await this.dbClient.query(addQuery)
+    } catch (error) {
+      this.catchDataBaseErrror(error)
+    }
+  }
+
+  async modifyEntrey(oldObject: any, newObject: any, tableName: TableName): Promise<void> {
+
+  }
+
+  catchDataBaseErrror(error: any) {
+    if (error instanceof DatabaseError) {
+      throw new API_Error('DATABASE_ERROR', `Erreur base de donnée : ${error.detail}`, {code: 400})
+    } else {
+      console.log(error)
+      throw new API_Error('API_INTERN_ERROR', 'Une erreur interne est survenue...')
     }
   }
 
